pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = "A Django blog application"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2"
django-ninja = "^1.4.1"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
dj-database-url = "^2.3.0"
python-json-logger = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
